Rafeys changes

+++ UniversityCourseSelection/src/main/java/com/capgemini/UniversityCourseSelection/controllers/ApplicantController.java
@@ -2,6 +2,7 @@ package com.capgemini.UniversityCourseSelection.controllers;
 
 
 import com.capgemini.UniversityCourseSelection.entities.Applicant;
+import com.capgemini.UniversityCourseSelection.exception.NotFoundException;
 import com.capgemini.UniversityCourseSelection.repo.ILoginRepository;
 import com.capgemini.UniversityCourseSelection.services.IApplicantService;
 
@@ -15,12 +16,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Optional;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/applicant")
@@ -31,6 +33,25 @@ public class ApplicantController {
 //	
 //	@Autowired
 //	private ILoginRepository repo;
+	
+	private boolean checkSession(HttpServletRequest request,String type) {
+		HttpSession session =  request.getSession();
+
+		boolean validLogin = true;
+		if(session.isNew()) {
+			validLogin = false;
+		}
+		else {
+			int userId = (int) session.getAttribute(type);
+            if(userId!=0)
+            	validLogin=true;
+		}
+//		if(invalidLogin)
+//			
+		return validLogin;
+	}
+
+	
 
 	@PostMapping("/apply")
 	public ResponseEntity<Applicant> applyForCourse(@RequestBody Applicant applicant) {
@@ -43,6 +64,9 @@ public class ApplicantController {
 
 	@PutMapping("/update")
 	public ResponseEntity<Applicant> updateApplication(@RequestBody Applicant applicant) {
+		if(applicant==null || applicant.getApplicantId()==null) {
+			throw new NotFoundException("Applicant or Id can't be null!");
+		}
 
 		Applicant temp = service.updateApplicant(applicant);
 
@@ -51,7 +75,16 @@ public class ApplicantController {
 	}
 
 	@DeleteMapping("/delete")
-	public ResponseEntity<Applicant> deleteApplication(@RequestBody Applicant applicant) {
+	public ResponseEntity<Applicant> deleteApplication(@RequestBody Applicant applicant,HttpServletRequest request) {
+		 boolean valid=checkSession(request,"commitee");
+			
+			if(!valid) {
+				return new ResponseEntity("Please Login to Access <a href = 'localhost:8080/login/commitee'>Login</a> ",HttpStatus.FORBIDDEN);
+				
+			}
+		if(applicant==null || applicant.getApplicantId()==null) {
+			throw new NotFoundException("Applicant or Id can't be null!");
+		}
 		Applicant temp = service.deleteApplicant(applicant);
 		return new ResponseEntity<>(temp, HttpStatus.OK);
 
@@ -60,6 +93,7 @@ public class ApplicantController {
 	
 	@GetMapping("/get/{id}")
 	public ResponseEntity<Applicant> getById(@PathVariable("id") Integer id){
+		
 		Optional<Applicant> temp= service.viewApplicant(id);
 		if(temp.isEmpty()) {
 			return new ResponseEntity("No such user",HttpStatus.NOT_FOUND);
@@ -69,7 +103,13 @@ public class ApplicantController {
 	}
 	
 	@GetMapping("/getAll/{status}")
-	public ResponseEntity<List<Applicant>> getAllApplicants(@PathVariable int status){
+	public ResponseEntity<List<Applicant>> getAllApplicants(@PathVariable int status,HttpServletRequest request){
+        boolean valid=checkSession(request,"commitee");
+		
+		if(!valid) {
+			return new ResponseEntity("Please Login to Access <a href = 'localhost:8080/login/commitee'>Login</a> ",HttpStatus.FORBIDDEN);
+			
+		}
 		
 		List<Applicant> res= service.viewAllApplicantsByStatus(status);
 		return new ResponseEntity<>(res,HttpStatus.OK);
